Clarify in the docs which object is the considered the "old" vs "new".
Some diff'ing functions, particularly for testing use the sematics of `diff(actual,expected)`. You see this for example in the `unfunk-diff` library uses this library. 

Other tools like this one use the sematics of `diff(old,new)`. Either can make sense, but it's not clear which semantics are being used when the example arguments are 'ObjectA' and 'ObjectB'.   

Updating the example input names clarify which semantics are being used here.